Clean up the unused viewdata models Refactor create non-concerns journey to enable back button - moved the partials into individual pages that can be redirected to
