+++ ConcernsCaseWork/ConcernsCaseWork/Constants/ViewDataConstants.cs
@@ -2,9 +2,7 @@
 {
 	public static class ViewDataConstants
 	{
-		public const string BackButtonLink = "BackButtonLink";
 		public const string Title = "Title";
-		public const string BackButtonLabel = "BackButtonLabel";
 
 	}
 }
