--- ConcernsCaseWork/ConcernsCaseWork/Pages/Case/CreateCase/NonConcernsCase/CreateNonConcernsCase.cshtml
@@ -1,38 +1,0 @@
-@page "/case/create/nonconcerns"
-@using Microsoft.AspNetCore.Mvc.TagHelpers
-@model CreateNonConcernsCasePageModel
-
-@{
-	ViewData["Title"] = "Add to Case";
-	if (Model.SelectedActionOrDecision == CreateNonConcernsCasePageModel.Options.None)
-	{
-		ViewData["BackButtonLabel"] = "Back to case overview";
-		ViewData["BackButtonLink"] = "/";
-	}
-	else
-	{
-		ViewData["BackButtonLink"] = "/case/create/nonconcerns";
-	}
-}
-
-@section BeforeMain {
-	<div class="govuk-extra-width-container">
-		<back-link url="@Model.BackLink.Url" label="@Model.BackLink.Label"></back-link>
-	</div>
-}
-
-<div class="govuk-width-container">
-	<partial name="_BannerError"/>
-
-	@if (Model.SelectedActionOrDecision == CreateNonConcernsCasePageModel.Options.None )
-	{
-		<partial name="_SelectActionOrDecision"/>
-	}
-	@if (Model.SelectedActionOrDecision == CreateNonConcernsCasePageModel.Options.Action )
-	{
-		<partial name="_SelectAction"/>
-	}
-
-</div>
-
-
