+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/CreateCase/NonConcernsCase/SelectAction.cshtml
@@ -1,0 +1,62 @@
+@page "/case/create/nonconcerns/action"
+@using Microsoft.AspNetCore.Mvc.TagHelpers
+@model SelectActionPageModel
+
+@{
+	ViewData["Title"] = "Add to Case";
+}
+
+@section BeforeMain {
+	<div class="govuk-width-container">
+		<back-link url="@Model.BackLink.Url" label="@Model.BackLink.Label"></back-link>
+	</div>
+}
+
+<div class="govuk-width-container" id="action">
+	
+	<partial name="_BannerError"/>
+	
+	@if (!string.IsNullOrEmpty((string)TempData.Peek("Error.Message")))
+	{
+		<partial name="_Error"/>
+	}
+	else
+	{
+		<form role="form" method="post" id="choose-action-form" novalidate>
+			<partial name="_ValidationErrors"/>
+			<h1 class="govuk-heading-l">Add to case</h1>
+
+			<h2 class="govuk-heading-m">Select an action</h2>
+
+			<div class="govuk-radios" data-module="govuk-radios">
+				<div class="govuk-radios__item">
+					<input asp-for="SelectedAction" class="govuk-radios__input" id="srma-option" type="radio" value="@SelectActionPageModel.Actions.SRMA">
+					<label class="govuk-label govuk-radios__label" for="srma-option">
+						School Management Resource Advisor (SRMA)
+					</label>
+				</div>
+				<div class="govuk-radios__item">
+					<input asp-for="SelectedAction" class="govuk-radios__input" id="tff-option" type="radio" value="@SelectActionPageModel.Actions.TFF" disabled>
+					<label class="govuk-label govuk-radios__label" for="tff-option">
+						Trust Financial Forecast
+					</label>
+				</div>
+			</div>
+
+			<div class="govuk-button-group">
+				<button data-prevent-double-click="true"
+				        class="govuk-button govuk-!-margin-top-6"
+				        data-module="govuk-button"
+				        role="button"
+				        id="continue"
+				        name="action"
+				        value="continue">
+					Continue
+				</button>
+			</div>
+
+		</form>
+	}
+	</div>
+
+
