+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/CreateCase/NonConcernsCase/SelectActionOrDecision.cshtml
@@ -1,9 +1,17 @@
+@page "/case/create/nonconcerns"
 @using Microsoft.AspNetCore.Mvc.TagHelpers
-@model CreateNonConcernsCasePageModel
+@model SelectActionOrDecisionPageModel
+
 @{
-	ViewData["Title"] = "Create case";
-	ViewData["BackButtonLink"] = "/case";
+	ViewData["Title"] = "Add to Case";
 }
+
+@section BeforeMain {
+	<div class="govuk-width-container">
+		<back-link url="@Model.BackLink.Url" label="@Model.BackLink.Label"></back-link>
+	</div>
+}
+
 <div class="govuk-width-container" id="actionordecision">
 	
 	<partial name="_BannerError"/>
@@ -18,19 +26,19 @@
 			<partial name="_ValidationErrors"/>
 			<h1 class="govuk-heading-l">Add to case</h1>
 
-			<h2 class="govuk-heading-m">Select an action</h2>
+			<h2 class="govuk-heading-m">Are you recording an action or decision?</h2>
 
 			<div class="govuk-radios" data-module="govuk-radios">
 				<div class="govuk-radios__item">
-					<input asp-for="SelectedAction" class="govuk-radios__input" id="srma-option" type="radio" value="@CreateNonConcernsCasePageModel.Actions.SRMA">
-					<label class="govuk-label govuk-radios__label" for="srma-option">
-						School Management Resource Advisor (SRMA)
+					<input asp-for="SelectedActionOrDecision" class="govuk-radios__input" id="action-option" type="radio" value="@SelectActionOrDecisionPageModel.Options.Action">
+					<label class="govuk-label govuk-radios__label" for="action-option">
+						Action
 					</label>
 				</div>
 				<div class="govuk-radios__item">
-					<input asp-for="SelectedAction" class="govuk-radios__input" id="tff-option" type="radio" value="@CreateNonConcernsCasePageModel.Actions.TFF" disabled>
-					<label class="govuk-label govuk-radios__label" for="tff-option">
-						Trust Financial Forecast
+					<input asp-for="SelectedActionOrDecision" class="govuk-radios__input" id="decision-option" type="radio" value="@SelectActionOrDecisionPageModel.Options.Decision">
+					<label class="govuk-label govuk-radios__label" for="decision-option">
+						Decision
 					</label>
 				</div>
 			</div>
@@ -50,3 +58,5 @@
 		</form>
 	}
 	</div>
+
+
