+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/CreateCase/SelectCaseType.cshtml.cs
@@ -1,46 +1,40 @@
 using Ardalis.GuardClauses;
 using ConcernsCaseWork.Authorization;
-using ConcernsCaseWork.Extensions;
-using ConcernsCaseWork.Helpers;
+using ConcernsCaseWork.Constants;
 using ConcernsCaseWork.Logging;
 using ConcernsCaseWork.Models;
 using ConcernsCaseWork.Pages.Base;
 using ConcernsCaseWork.Redis.Models;
 using ConcernsCaseWork.Redis.Users;
-using ConcernsCaseWork.Service.Trusts;
 using ConcernsCaseWork.Services.Trusts;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using System;
-using System.Collections.Generic;
-using System.Net;
 using System.Threading.Tasks;
 
 namespace ConcernsCaseWork.Pages.Case.CreateCase;
 
 [Authorize]
 [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-public class CreateCasePageModel : AbstractPageModel
+public class SelectCaseTypePageModel : AbstractPageModel
 {
 	private readonly ITrustModelService _trustModelService;
 	private readonly IUserStateCachedService _cachedUserService;
-	private readonly ILogger<CreateCasePageModel> _logger;
+	private readonly ILogger<SelectCaseTypePageModel> _logger;
 	private readonly IClaimsPrincipalHelper _claimsPrincipalHelper;
-	private const int _searchQueryMinLength = 3;
 
 	[BindProperty(SupportsGet = true)]
 	public TrustAddressModel TrustAddress { get; set; }
 		
 	[BindProperty]
 	public CaseTypes CaseType { get; set; }
+	
+	public Hyperlink BackLink => BuildBackLinkFromHistory(fallbackUrl: PageRoutes.YourCaseworkHomePage);
 
-	[TempData]
-	public CreateCaseSteps CreateCaseStep { get; set; } = CreateCaseSteps.SearchForTrust;
-
-	public CreateCasePageModel(ITrustModelService trustModelService,
+	public SelectCaseTypePageModel(ITrustModelService trustModelService,
 		IUserStateCachedService cachedUserService,
-		ILogger<CreateCasePageModel> logger,
+		ILogger<SelectCaseTypePageModel> logger,
 		IClaimsPrincipalHelper claimsPrincipalHelper)
 	{
 		_trustModelService = Guard.Against.Null(trustModelService);
@@ -55,10 +49,7 @@ public class CreateCasePageModel : AbstractPageModel
 		
 		try
 		{
-			if (CreateCaseStep == CreateCaseSteps.SelectCaseType)
-			{
-				await SetTrustAddress();
-			}
+			await SetTrustAddress();
 		}
 		catch (Exception ex)
 		{
@@ -73,16 +64,10 @@ public class CreateCasePageModel : AbstractPageModel
 	public async Task<IActionResult> OnPost()
 	{
 		_logger.LogMethodEntered();
-			
+		
 		try
 		{
-			if (CreateCaseStep != CreateCaseSteps.SelectCaseType)
-			{
-				throw new Exception();
-			}
-
 			await ResetUserState();
-			ResetCurrentStep();
 
 			switch (CaseType)
 			{
@@ -111,77 +96,8 @@ public class CreateCasePageModel : AbstractPageModel
 			userState.CreateCaseModel = new CreateCaseModel();
 			await _cachedUserService.StoreData(userName, userState);
 		}
-		
-		void ResetCurrentStep()
-		{
-			CreateCaseStep = CreateCaseSteps.SearchForTrust;
-		}
 	}
 	
-	// This is an AJAX call
-	public async Task<IActionResult> OnGetTrustsSearchResult(string searchQuery)
-	{
-		_logger.LogMethodEntered();
-		
-		try
-		{
-			if (!SearchQueryIsValid()) 
-				return new JsonResult(new List<TrustSearchModel>());
-
-			var trustSearchResponse = await BuildTrustResponse();
-			
-			return new JsonResult(trustSearchResponse);
-		}
-		catch (Exception ex)
-		{
-			return HandleExceptionForAjaxCall(ex);
-		}
-		
-		bool SearchQueryIsValid() => !(string.IsNullOrEmpty(searchQuery) || searchQuery.Length < _searchQueryMinLength);
-
-		async Task<IList<TrustSearchModel>> BuildTrustResponse()
-		{
-			var trustSearch = new TrustSearch(searchQuery, searchQuery, searchQuery);
-			return await _trustModelService.GetTrustsBySearchCriteria(trustSearch);
-		}
-	}
-
-	// This is an AJAX call
-	public async Task<IActionResult> OnGetSelectedTrust(string trustUkPrn)
-	{
-		_logger.LogMethodEntered();
-		
-		try
-		{
-			if (!TrustUkPrnIsValid())
-				throw new Exception($"Selected trust is incorrect - {trustUkPrn}");
-
-			await CacheTrustUkPrn();
-			
-			SetNextStep();
-
-			return ReloadPageForNextStep();
-		}
-		catch (Exception ex)
-		{
-			return HandleExceptionForAjaxCall(ex);
-		}
-		
-		bool TrustUkPrnIsValid() => !(string.IsNullOrEmpty(trustUkPrn) || trustUkPrn.Contains('-') || trustUkPrn.Length < _searchQueryMinLength);
-
-		async Task CacheTrustUkPrn()
-		{
-			var userName = GetUserName();
-			var userState = await _cachedUserService.GetData(userName);
-			userState.TrustUkPrn = trustUkPrn;
-			await _cachedUserService.StoreData(userName, userState);
-		}
-		
-		void SetNextStep() => CreateCaseStep = CreateCaseSteps.SelectCaseType;
-		
-		JsonResult ReloadPageForNextStep() => new (new { redirectUrl = "/case/create" });
-	}
-
 	private async Task SetTrustAddress()
 	{
 		var userName = GetUserName();
@@ -201,13 +117,6 @@ public class CreateCasePageModel : AbstractPageModel
 		TrustAddress = trustAddress ?? throw new Exception($"Could not find trust with UK PRN of {userState.TrustUkPrn}");
 	}
 
-	private ObjectResult HandleExceptionForAjaxCall(Exception ex)
-	{
-		_logger.LogErrorMsg(ex);
-			
-		return StatusCode((int)HttpStatusCode.InternalServerError, ex.Message);
-	}
-
 	private string GetUserName() => _claimsPrincipalHelper.GetPrincipalName(User);
 
 	public enum CaseTypes
@@ -215,11 +124,5 @@ public class CreateCasePageModel : AbstractPageModel
 		NotSelected,
 		Concern,
 		NonConcern
-	}
-	
-	public enum CreateCaseSteps
-	{
-		SearchForTrust,
-		SelectCaseType
 	}
 }
