+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/Index.cshtml
@@ -5,7 +5,6 @@
 
 @{
 	ViewData["Title"] = "New Case";
-	ViewData["BackButtonLink"] = "/";
 	var nonce = HttpContext.GetNonce();
 }
