+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/Management/Action/Decision/Close.cshtml
@@ -5,14 +5,12 @@
 
 @{
 	ViewData["Title"] = "Close decision";
-	ViewData["BackButtonLabel"] = "Back to case";
-	ViewData["BackButtonLink"] = $"/case/{@Model.CaseUrn}/management";
 	var nonce = HttpContext.GetNonce();
 }
 
 @section BeforeMain {
 	<div class="govuk-width-container">
-		<back-link url="@Model.BackLink.Url" label="@Model.BackLink.Label"></back-link>
+		<back-link url="/case/@RouteData.Values["urn"]/management" label="Back to case overview"></back-link>
 	</div>
 }
