+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/Management/Action/Decision/Index.cshtml
@@ -5,7 +5,7 @@
 
 @section BeforeMain {
 	<div class="govuk-width-container">
-		<back-link url="@Model.BackLink.Url" label="@Model.BackLink.Label"></back-link>
+		<back-link url="/case/@RouteData.Values["urn"]/management" label="Back to case overview"></back-link>
 	</div>
 }
