+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/Management/Action/Decision/Index.cshtml.cs
@@ -17,8 +17,6 @@ namespace ConcernsCaseWork.Pages.Case.Management.Action.Decision
 		private ILogger _logger;
 
 		public ViewDecisionModel Decision { get; set; }
-		
-		public Hyperlink BackLink => BuildBackLinkFromHistory(fallbackUrl: PageRoutes.YourCaseworkHomePage, label: "Back to case overview");
 
 		public IndexPageModel(
 			IDecisionService decisionService,
@@ -33,7 +31,6 @@ namespace ConcernsCaseWork.Pages.Case.Management.Action.Decision
 			_logger.LogInformation("Case::Action::Decision::IndexPageModel::OnGetAsync");
 
 			ViewData[ViewDataConstants.Title] = "Decision";
-			ViewData[ViewDataConstants.BackButtonLabel] = "Back to case overview";
 
 			try
 			{
@@ -43,8 +40,6 @@ namespace ConcernsCaseWork.Pages.Case.Management.Action.Decision
 				var apiDecision = await _decisionService.GetDecision(urn, (int)decisionId);
 
 				Decision = DecisionMapping.ToViewDecisionModel(apiDecision);
-
-				ViewData[ViewDataConstants.BackButtonLink] = Decision.BackLink;
 			}
 			catch (Exception ex)
 			{
