+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/ViewClosed.cshtml
@@ -11,7 +11,7 @@
 }
 
 @section BeforeMain {
-	<div class="govuk-extra-width-container">
+	<div class="govuk-width-container">
 		<back-link url="@Model.BackLink.Url" label="@Model.BackLink.Label"></back-link>
 	</div>
 }
